DO NOT MERGE Crude welsh only implementation
This is an implementation of a "only_welsh" permission that restricts
access to users with this permission so that they can only access the
Welsh content in Publisher.

This relies on hiding things for the purpose of a demo and would leave
any enterprising user with this permission with the ability to do a fair
amount of stuff they shouldn't.

Inherently it's a bad idea to use a permission as a restriction. Instead
it would make more sense to have a gds_editor permission which has more
access than a welsh permission.